Move EntitiesContent
